requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
attrs==25.1.0
base58==2.1.1
borsh-construct==0.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
construct==2.10.68
construct-typing==0.5.6
frozenlist==1.5.0
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonalias==0.1.1
multidict==6.1.0
propcache==0.2.1
protobuf==5.29.3
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
solana==0.36.3
solders==0.23.0
sumtypes==0.1a6
typing_extensions==4.12.2
urllib3==2.3.0
websockets==13.1
yarl==1.18.3
